requirements.txt
Django==3.2
backports.zoneinfo==0.2.1
graphene-django==3.1.1
graphql-core==3.2.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-cors-headers==3.10.0
django-graphql-jwt==0.3.4
celery==5.3.0
redis==4.5.0
todoist-api-python==2.0.2